+++ packages/terra-framework-docs/CHANGELOG.md
@@ -9,6 +9,7 @@
 * Changed
   * Updated examples and tests for `terra-table` to consume updated onRowSelect callback.
   * Updated XFC initialization for the Embedded Content Consumer examples.
+  * Updated Flowsheet Data Grid Sections Example to showcase multi-select and range selection on sections. 
 
 ## 1.49.0 - (December 1, 2023)